pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "imap-client>=0.1.8",
    "imap-tools>=1.11.1",
    "langchain>=1.2.10",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.9",
    "python-dotenv>=1.2.1",
]
